Handle QR chunks for SigningResponse on cold wallet side #15
