+++ app/src/main/res/layout/fragment_cold_wallet_signing.xml
@@ -144,21 +144,31 @@
                     android:layout_height="wrap_content"
                     android:orientation="vertical">
 
-                    <ImageView
-                        android:id="@+id/qr_code"
+                    <androidx.viewpager2.widget.ViewPager2
+                        android:id="@+id/qr_code_pager"
                         android:layout_width="match_parent"
-                        android:layout_height="325dp"
-                        android:layout_marginVertical="@dimen/activity_vertical_margin"
-                        android:contentDescription="@string/label_qr_code"
-                        tools:src="@drawable/ic_qr_code_24" />
+                        android:layout_height="wrap_content"
+                        tools:layout_height="325dp" />
 
                     <TextView
+                        android:id="@+id/tv_scan_signed_desc"
                         style="@style/TextAppearance.App.Body1"
                         android:layout_width="match_parent"
                         android:layout_height="wrap_content"
                         android:layout_marginHorizontal="@dimen/activity_horizontal_margin"
                         android:gravity="center"
                         android:text="@string/desc_show_signed" />
+
+                    <Button
+                        android:id="@+id/button_scan_next_qr"
+                        style="?attr/materialButtonStyle"
+                        android:layout_width="wrap_content"
+                        android:layout_height="wrap_content"
+                        android:layout_gravity="center_horizontal"
+                        android:layout_marginTop="@dimen/activity_horizontal_margin"
+                        android:layout_marginBottom="24dp"
+                        android:text="@string/button_next"
+                        android:visibility="gone" />
 
                     <Button
                         android:id="@+id/button_dismiss"
@@ -180,7 +190,7 @@
                 android:layout_gravity="center"
                 android:layout_margin="@dimen/activity_horizontal_margin"
                 android:visibility="gone"
-                tools:visibility="visible">
+                tools:visibility="gone">
 
                 <LinearLayout
                     android:layout_width="match_parent"
